feat(numbers): RFC 8785 §3.2.2.3 float_to_str; fix float literals in examples/CI
Steps 2-5 of [capsule-emit-float-in-digest-example-broken]:

- Add capsule_emit/numbers.py: float_to_str() implements ES6 Number::toString
  (ECMA-262 §7.1.12.1 steps 6-10), RFC 8785 Appendix B (-0.0 → "0"),
  raises FloatInDigestError for NaN/Inf. ONE RULE ONE PLACE.
  PROVISIONAL — rule sent to Anton 2026-08-17; confirm expected at VTO 2026-08-19.

- Export float_to_str from capsule_emit.__init__ under "Number conversion" comment.

- Fix examples/quickstart_demo.py: "total": 1240.19 → "total": "1240.19"

- Fix examples/wicket/demo.py: amount: float → int at 3 function signatures
  and 3 call sites (AmountUnderCap.check does numeric comparison; str would TypeError).

- Extend tests/test_examples_smoke.py: add test_wicket_demo() and
  test_quickstart_demo_offline() (AAC_ANCHOR_URL=127.0.0.1:19999 for fast offline).

- Add tests/test_float_to_str.py: 33 vectors covering all ECMA-262 §7.1.12.1
  branches, -0.0 → "0", NaN/Inf raises, round-trip property, float32-widening.

+++ examples/wicket/demo.py
@@ -155,11 +155,11 @@ mcp_emitter = MCPCapsuleEmitter(
 
 
 @mcp_emitter.tool(constraints=[AmountUnderCap(5000), VendorKnown({"Acme", "Globex"})])
-def submit_order(vendor: str, amount: float) -> dict:
+def submit_order(vendor: str, amount: int) -> dict:
     return {"status": "accepted", "vendor": vendor, "amount": amount}
 
 
-mcp_result = submit_order(vendor="Globex", amount=2500.0)
+mcp_result = submit_order(vendor="Globex", amount=2500)
 print(f"tool returned: {mcp_result}")
 
 assert mcp_emitter.last is not None
@@ -190,11 +190,11 @@ mcp_block_calls: list = []
     constraints=[AmountUnderCap(100)],
     on_block=lambda action, gr: mcp_block_calls.append((action, gr)),
 )
-def blocked_order(vendor: str, amount: float) -> dict:
+def blocked_order(vendor: str, amount: int) -> dict:
     return {"status": "pending"}
 
 
-blocked_result = blocked_order(vendor="Acme", amount=999.0)
+blocked_result = blocked_order(vendor="Acme", amount=999)
 print(f"tool returned: {blocked_result}")
 
 assert mcp_block_emitter.last is not None
@@ -217,12 +217,12 @@ err_emitter = MCPCapsuleEmitter(
 
 
 @err_emitter.tool(constraints=[AmountUnderCap(100)])
-def order_no_callback(vendor: str, amount: float) -> dict:
+def order_no_callback(vendor: str, amount: int) -> dict:
     return {"status": "ok"}
 
 
 try:
-    order_no_callback(vendor="Acme", amount=9999.0)
+    order_no_callback(vendor="Acme", amount=9999)
     print("ERROR: should have raised GateBlockedError")
     sys.exit(1)
 except GateBlockedError as e:
